chore: extract period validation rules into shared module
Moves the period-tab validation rules (period type, start/end dates,
Timeline/Split min-max) out of validateThematicLayer.js into
components/edit/shared/validatePeriod.js, and reuses them in
EventDialog's own validate(), replacing its ad hoc period/date check
and adding the "Period is required" rule that multi-select presets
now need. renderingStrategy stays optional so Events (no rendering
strategy yet) simply never trigger the Timeline/Split-only rules.
